@@ -1,5 +1,5 @@
 {
-  "private": true,
+  "type": "module",
   "packageManager": "pnpm@8.9.2",
   "scripts": {
     "build": "vite build",
@@ -12,41 +12,41 @@
     "up": "taze major -I"
   },
   "dependencies": {
-    "@element-plus/icons-vue": "^2.1.0",
-    "@vueuse/core": "^10.5.0",
-    "axios": "^0.21.4",
-    "element-plus": "^2.4.1",
+    "@element-plus/icons-vue": "^2.3.1",
+    "@vueuse/core": "^10.8.0",
+    "axios": "^1.6.7",
+    "element-plus": "^2.5.6",
     "lodash-es": "^4.17.21",
-    "vue": "^3.3.4",
-    "vue-router": "^4.2.5"
+    "vue": "^3.4.20",
+    "vue-router": "^4.3.0"
   },
   "devDependencies": {
     "@antfu/eslint-config": "^2.6.4",
-    "@iconify-json/carbon": "^1.1.21",
-    "@types/lodash-es": "^4.17.10",
-    "@types/mockjs": "^1.0.9",
-    "@types/node": "^18.18.6",
-    "@unocss/eslint-config": "^0.56.5",
+    "@iconify-json/carbon": "^1.1.30",
+    "@types/lodash-es": "^4.17.12",
+    "@types/mockjs": "^1.0.10",
+    "@types/node": "^18.19.18",
     "@unocss/eslint-plugin": "^0.58.5",
-    "@unocss/reset": "^0.56.5",
-    "@unocss/transformer-attributify-jsx": "^0.56.5",
-    "@vitejs/plugin-vue": "^4.4.0",
-    "@vitejs/plugin-vue-jsx": "^3.0.2",
-    "@vue/test-utils": "^2.4.1",
+    "@unocss/reset": "^0.58.5",
+    "@unocss/transformer-attributify-jsx": "^0.58.5",
+    "@vitejs/plugin-vue": "^4.6.2",
+    "@vitejs/plugin-vue-jsx": "^3.1.0",
+    "@vue/test-utils": "^2.4.4",
     "eslint": "^8.57.0",
     "jsdom": "^22.1.0",
     "mockjs": "^1.1.0",
-    "pnpm": "^8.9.2",
+    "pnpm": "^8.15.4",
+    "rollup-plugin-visualizer": "^5.12.0",
     "taze": "^0.11.4",
-    "typescript": "^5.2.2",
-    "unocss": "^0.56.5",
-    "unplugin-auto-import": "^0.16.6",
-    "unplugin-vue-components": "^0.25.2",
-    "vite": "^4.5.0",
-    "vite-plugin-mock": "^2.9.6",
-    "vite-plugin-pages": "^0.31.0",
-    "vite-plugin-vue-devtools": "^0.5.1",
-    "vitest": "^0.34.6",
-    "vue-tsc": "^1.8.19"
+    "typescript": "^5.3.3",
+    "unocss": "^0.58.5",
+    "unplugin-auto-import": "^0.17.5",
+    "unplugin-vue-components": "^0.26.0",
+    "vite": "^5.1.4",
+    "vite-plugin-mock": "^3.0.1",
+    "vite-plugin-pages": "^0.32.0",
+    "vite-plugin-vue-devtools": "^7.0.15",
+    "vitest": "^1.3.1",
+    "vue-tsc": "^1.8.27"
   }
 }
